% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdVignettes-package.R
\docType{package}
\name{mdVignettes}
\alias{mdVignettes}
\alias{mdVignettes-package}
\title{mdVignettes}
\description{
mdVignettes provides a vignette engine for markdown vignettes.
}
\details{
To use the vignette engine, add

\preformatted{
    Suggests: mdVignettes
    VignetteBuilder: mdVignettes
}

to your DESCRIPTION file, and add YAML front matter:

\preformatted{
    ---
    title: "<Your Vignette Title Here>"
    author: "<Your Name Here>"
    date: "<Date Here>"
    output: html_document
    vignette: >
      \%\VignetteIndexEntry{<Your Vignette Title Here>}
      \%\VignetteEngine{mdVignettes::md}
      \%\VignetteEncoding{UTF-8}
    ---
}

If you would prefer to include your own custom CSS stylesheet
(rather than use the default stylesheet from the \code{markdown} package),
you can do the following:

\preformatted{
    ---
    title: "<Your Vignette Title Here>"
    author: "<Your Name Here>"
    date: "<Date Here>"
    output:
      html_document:
        css: <Your stylesheet filename here>
    vignette: >
      \%\VignetteIndexEntry{<Your Vignette Title Here>}
      \%\VignetteEngine{mdVignettes::md}
      \%\VignetteEncoding{UTF-8}
    ---
}
}
\author{
JB Duck-Mayr
}
